mtd: nand: Kconfig Setup missing NAND dependencies based on ARCH
To avoid clutter in the NAND menu, let's make the NAND controllers
dependent on which ARCH type uses them.  This way, we cannot
accidentally try to enable a NAND controller that doesn't exist.

This is based on Convert CONFIG_SYS_MAX_NAND_DEVICE to Kconfig and
add CONFIG_NAND that was just submitted.

Signed-off-by: Adam Ford <aford173@gmail.com>